Changed type to u32

+++ runtime/storage/arrays/StoredU128Array.ts
@@ -9,7 +9,7 @@ import { bigEndianAdd } from '../../math/bytes';
  */
 @final
 export class StoredU128Array extends StoredPackedArray<u128> {
-    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u64 = DEFAULT_MAX_LENGTH) {
+    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u32 = DEFAULT_MAX_LENGTH) {
         super(pointer, subPointer, u128.Zero, maxLength);
     }
