+++ runtime/storage/arrays/StoredU16Array.ts
@@ -6,7 +6,7 @@ import { bigEndianAdd } from '../../math/bytes';
  */
 @final
 export class StoredU16Array extends StoredPackedArray<u16> {
-    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u64 = DEFAULT_MAX_LENGTH) {
+    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u32 = DEFAULT_MAX_LENGTH) {
         super(pointer, subPointer, 0, maxLength);
     }
