+++ runtime/storage/arrays/StoredU256Array.ts
@@ -8,7 +8,7 @@ import { bigEndianAdd } from '../../math/bytes';
  */
 @final
 export class StoredU256Array extends StoredPackedArray<u256> {
-    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u64 = DEFAULT_MAX_LENGTH) {
+    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u32 = DEFAULT_MAX_LENGTH) {
         super(pointer, subPointer, u256.Zero, maxLength);
     }
