+++ runtime/storage/arrays/StoredU32Array.ts
@@ -7,7 +7,7 @@ import { bigEndianAdd } from '../../math/bytes';
  */
 @final
 export class StoredU32Array extends StoredPackedArray<u32> {
-    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u64 = DEFAULT_MAX_LENGTH) {
+    public constructor(pointer: u16, subPointer: Uint8Array, maxLength: u32 = DEFAULT_MAX_LENGTH) {
         super(pointer, subPointer, 0, maxLength);
     }
